feat(directives): add new directives
new directives:
<ghanoz-countries-typeahead></ghanoz-countries-typeahead>
<ghanoz-countries-select></ghanoz-countries-select>

we drop the external template to reduce headache when deal with the path url. Since the template is only has a few lines of codes, it does make sense to inline the template.

the current issue we had now is I need to find out how to let the directive pass model with custom name or any name user like.